user_profile/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0